# frontend/src/_generated_code/openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/v1/login/access-token":{"post":{"tags":["login"],"summary":"Login Access Token","description":"OAuth2 compatible token login, get an access token for future requests\n\nRaises: Http404UserNotFound, Http400IncorrectPassword","operationId":"login_access_token","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_access_token_api_v1_login_access_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/quizzes/create":{"post":{"tags":["quizzes"],"summary":"Create Quiz","operationId":"create_quiz","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Quiz"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/quizzes/getPreviewsByUser":{"get":{"tags":["quizzes"],"summary":"Get Quiz Preview By User","operationId":"get_quiz_preview_by_user","parameters":[{"required":true,"schema":{"title":"Owner Id","type":"string","format":"uuid"},"name":"owner_id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Quiz Preview By User Api V1 Quizzes Getpreviewsbyuser Get","type":"array","items":{"$ref":"#/components/schemas/Quiz"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/quizzes/{id}":{"get":{"tags":["quizzes"],"summary":"Get Quiz","operationId":"get_quiz","parameters":[{"required":true,"schema":{"title":"Id","type":"string","format":"uuid"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Quiz"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"delete":{"tags":["quizzes"],"summary":"Delete Quiz","operationId":"delete_quiz","parameters":[{"required":true,"schema":{"title":"Id","type":"string","format":"uuid"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/quizzes/for_student/{id}":{"get":{"tags":["quizzes"],"summary":"Get Quiz For Student","operationId":"get_quiz_for_student","parameters":[{"required":true,"schema":{"title":"Id","type":"string","format":"uuid"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizForStudent"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/quizzes/update":{"patch":{"tags":["quizzes"],"summary":"Update Quiz","operationId":"update_quiz","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Quiz"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/quizzes/question/create":{"post":{"tags":["quizzes"],"summary":"Create Question","operationId":"create_question","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Question"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/quizzes/question/update":{"patch":{"tags":["quizzes"],"summary":"Update Question","operationId":"update_question","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Question"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/quizzes/question/{id}":{"delete":{"tags":["quizzes"],"summary":"Delete Question","operationId":"delete_question","parameters":[{"required":true,"schema":{"title":"Id","type":"string","format":"uuid"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/quizzes/answer/create":{"post":{"tags":["quizzes"],"summary":"Create Answer","operationId":"create_answer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnswerCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Answer"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/quizzes/answer/update":{"patch":{"tags":["quizzes"],"summary":"Update Answer","operationId":"update_answer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnswerUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Answer"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/quizzes/answer/{id}":{"delete":{"tags":["quizzes"],"summary":"Delete Answer","operationId":"delete_answer","parameters":[{"required":true,"schema":{"title":"Id","type":"string","format":"uuid"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/rooms/create":{"post":{"tags":["rooms"],"summary":"Create Room","operationId":"create_room","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Room"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/rooms/get-by-code":{"get":{"tags":["rooms"],"summary":"Get Room By Code","operationId":"get_room_by_code","parameters":[{"required":true,"schema":{"title":"Code","type":"string"},"name":"code","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Room"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/rooms/update":{"put":{"tags":["rooms"],"summary":"Update Room","operationId":"update_room","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Room"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/users/create":{"post":{"tags":["users"],"summary":"Create User","operationId":"create_user","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/users/get-by-name":{"get":{"tags":["users"],"summary":"Get User By Name","operationId":"get_user_by_name","parameters":[{"required":true,"schema":{"title":"Name","type":"string"},"name":"name","in":"query"},{"required":true,"schema":{"title":"Number","type":"integer"},"name":"number","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/users/me":{"get":{"tags":["users"],"summary":"Read Current User","operationId":"read_current_user","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/sessions/create":{"post":{"tags":["sessions"],"summary":"Create Session","operationId":"create_session","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizSessionCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizSession"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/sessions/update":{"patch":{"tags":["sessions"],"summary":"Update Session","operationId":"update_session","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizSessionUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizSession"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/sessions/response/create":{"post":{"tags":["sessions"],"summary":"Create Student Response","operationId":"create_student_response","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentResponseCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/sessions/response/get_by_session":{"get":{"tags":["sessions"],"summary":"Get Responses By Session","operationId":"get_responses_by_session","parameters":[{"required":true,"schema":{"title":"Session Id","type":"string","format":"uuid"},"name":"session_id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"Answer":{"title":"Answer","required":["id","text","question_id","is_correct"],"type":"object","properties":{"previous_answer":{"title":"Previous Answer","type":"string","format":"uuid"},"id":{"title":"Id","type":"string","format":"uuid"},"text":{"title":"Text","type":"string"},"question_id":{"title":"Question Id","type":"string","format":"uuid"},"is_correct":{"title":"Is Correct","type":"boolean"}}},"AnswerCreate":{"title":"AnswerCreate","required":["text","question_id"],"type":"object","properties":{"previous_answer":{"title":"Previous Answer","type":"string","format":"uuid"},"text":{"title":"Text","type":"string"},"question_id":{"title":"Question Id","type":"string","format":"uuid"},"is_correct":{"title":"Is Correct","type":"boolean","default":false}}},"AnswerForStudent":{"title":"AnswerForStudent","required":["id","text","question_id"],"type":"object","properties":{"previous_answer":{"title":"Previous Answer","type":"string","format":"uuid"},"id":{"title":"Id","type":"string","format":"uuid"},"text":{"title":"Text","type":"string"},"question_id":{"title":"Question Id","type":"string","format":"uuid"}}},"AnswerUpdate":{"title":"AnswerUpdate","required":["id"],"type":"object","properties":{"previous_answer":{"title":"Previous Answer","type":"string","format":"uuid"},"id":{"title":"Id","type":"string","format":"uuid"},"text":{"title":"Text","type":"string"},"is_correct":{"title":"Is Correct","type":"boolean"}}},"Body_login_access_token_api_v1_login_access_token_post":{"title":"Body_login_access_token_api_v1_login_access_token_post","required":["username","password"],"type":"object","properties":{"grant_type":{"title":"Grant Type","pattern":"password","type":"string"},"username":{"title":"Username","type":"string"},"password":{"title":"Password","type":"string"},"scope":{"title":"Scope","type":"string","default":""},"client_id":{"title":"Client Id","type":"string"},"client_secret":{"title":"Client Secret","type":"string"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Question":{"title":"Question","required":["query","quiz_id","answers"],"type":"object","properties":{"previous_question":{"title":"Previous Question","type":"string","format":"uuid"},"id":{"title":"Id","type":"string","format":"uuid"},"query":{"title":"Query","type":"string"},"quiz_id":{"title":"Quiz Id","type":"string","format":"uuid"},"answers":{"title":"Answers","type":"array","items":{"$ref":"#/components/schemas/Answer"}}}},"QuestionCreate":{"title":"QuestionCreate","required":["query","quiz_id"],"type":"object","properties":{"previous_question":{"title":"Previous Question","type":"string","format":"uuid"},"query":{"title":"Query","type":"string"},"quiz_id":{"title":"Quiz Id","type":"string","format":"uuid"}}},"QuestionForStudent":{"title":"QuestionForStudent","required":["query","quiz_id","answers"],"type":"object","properties":{"previous_question":{"title":"Previous Question","type":"string","format":"uuid"},"id":{"title":"Id","type":"string","format":"uuid"},"query":{"title":"Query","type":"string"},"quiz_id":{"title":"Quiz Id","type":"string","format":"uuid"},"answers":{"title":"Answers","type":"array","items":{"$ref":"#/components/schemas/AnswerForStudent"}}}},"QuestionUpdate":{"title":"QuestionUpdate","required":["id"],"type":"object","properties":{"previous_question":{"title":"Previous Question","type":"string","format":"uuid"},"id":{"title":"Id","type":"string","format":"uuid"},"query":{"title":"Query","type":"string"}}},"Quiz":{"title":"Quiz","required":["id","title","owner_id","questions"],"type":"object","properties":{"id":{"title":"Id","type":"string","format":"uuid"},"title":{"title":"Title","type":"string"},"owner_id":{"title":"Owner Id","type":"string","format":"uuid"},"questions":{"title":"Questions","type":"array","items":{"$ref":"#/components/schemas/Question"}}}},"QuizCreate":{"title":"QuizCreate","required":["title","owner_id"],"type":"object","properties":{"title":{"title":"Title","type":"string"},"owner_id":{"title":"Owner Id","type":"string","format":"uuid"}}},"QuizForStudent":{"title":"QuizForStudent","required":["id","title","owner_id","questions"],"type":"object","properties":{"id":{"title":"Id","type":"string","format":"uuid"},"title":{"title":"Title","type":"string"},"owner_id":{"title":"Owner Id","type":"string","format":"uuid"},"questions":{"title":"Questions","type":"array","items":{"$ref":"#/components/schemas/QuestionForStudent"}}}},"QuizSession":{"title":"QuizSession","required":["id","room_id","quiz_id","is_active"],"type":"object","properties":{"id":{"title":"Id","type":"string","format":"uuid"},"room_id":{"title":"Room Id","type":"string","format":"uuid"},"quiz_id":{"title":"Quiz Id","type":"string","format":"uuid"},"is_active":{"title":"Is Active","type":"boolean"}}},"QuizSessionCreate":{"title":"QuizSessionCreate","required":["room_id","quiz_id"],"type":"object","properties":{"room_id":{"title":"Room Id","type":"string","format":"uuid"},"quiz_id":{"title":"Quiz Id","type":"string","format":"uuid"},"is_active":{"title":"Is Active","type":"boolean","default":true}}},"QuizSessionUpdate":{"title":"QuizSessionUpdate","required":["id"],"type":"object","properties":{"id":{"title":"Id","type":"string","format":"uuid"},"is_active":{"title":"Is Active","type":"boolean"}}},"QuizUpdate":{"title":"QuizUpdate","required":["id"],"type":"object","properties":{"id":{"title":"Id","type":"string","format":"uuid"},"title":{"title":"Title","type":"string"},"owner_id":{"title":"Owner Id","type":"string","format":"uuid"}}},"Room":{"title":"Room","required":["id","owner_id","code","is_active","session"],"type":"object","properties":{"id":{"title":"Id","type":"string","format":"uuid"},"owner_id":{"title":"Owner Id","type":"string","format":"uuid"},"code":{"title":"Code","type":"string"},"is_active":{"title":"Is Active","type":"boolean"},"session":{"title":"Session","type":"array","items":{"$ref":"#/components/schemas/QuizSession"}}}},"RoomCreate":{"title":"RoomCreate","required":["owner_id"],"type":"object","properties":{"owner_id":{"title":"Owner Id","type":"string","format":"uuid"},"code":{"title":"Code","type":"string"},"is_active":{"title":"Is Active","type":"boolean","default":true}}},"RoomUpdate":{"title":"RoomUpdate","required":["id"],"type":"object","properties":{"id":{"title":"Id","type":"string","format":"uuid"},"code":{"title":"Code","type":"string"},"is_active":{"title":"Is Active","type":"boolean"}}},"StudentResponse":{"title":"StudentResponse","required":["quiz_session_id","student_id","question_id","is_correct","ts"],"type":"object","properties":{"quiz_session_id":{"title":"Quiz Session Id","type":"string","format":"uuid"},"student_id":{"title":"Student Id","type":"string","format":"uuid"},"question_id":{"title":"Question Id","type":"string","format":"uuid"},"is_correct":{"title":"Is Correct","type":"boolean"},"ts":{"title":"Ts","type":"string","format":"date-time"}}},"StudentResponseCreate":{"title":"StudentResponseCreate","required":["quiz_session_id","student_id","question_id","is_correct"],"type":"object","properties":{"quiz_session_id":{"title":"Quiz Session Id","type":"string","format":"uuid"},"student_id":{"title":"Student Id","type":"string","format":"uuid"},"question_id":{"title":"Question Id","type":"string","format":"uuid"},"is_correct":{"title":"Is Correct","type":"boolean"}}},"Token":{"title":"Token","required":["access_token","token_type"],"type":"object","properties":{"access_token":{"title":"Access Token","type":"string"},"token_type":{"title":"Token Type","type":"string"}}},"User":{"title":"User","required":["id","display_name","is_active","is_superuser","number","identifier_name"],"type":"object","properties":{"id":{"title":"Id","type":"string","format":"uuid"},"display_name":{"title":"Display Name","type":"string"},"is_active":{"title":"Is Active","type":"boolean"},"is_superuser":{"title":"Is Superuser","type":"boolean"},"number":{"title":"Number","type":"integer"},"identifier_name":{"title":"Identifier Name","type":"string"}}},"UserCreate":{"title":"UserCreate","required":["display_name","password"],"type":"object","properties":{"display_name":{"title":"Display Name","type":"string"},"password":{"title":"Password","type":"string"},"is_active":{"title":"Is Active","type":"boolean","default":true},"is_superuser":{"title":"Is Superuser","type":"boolean","default":false}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/v1/login/access-token"}}}}}}